move select kernel command to extHost Kernel.
